Fixed slow response time. Added help menu to display available commands.
When the send call was outside of the switch statement it was being
executed every loop as fast as the arduino could handle which wasn’t
giving the led enough time to send a single message because it was
constantly starting a new one. This also explains why the wait time per
message varied from different commands.

Added a help menu to display commands.